+++ crates/miden-testing/src/kernel_tests/mod.rs
@@ -1,3 +1,3 @@
 pub(super) mod batch;
 mod block;
-pub(super) mod tx;
+pub(crate) mod tx;